+++ examples/example-app-router-without-i18n-routing/.eslintrc.js
@@ -5,6 +5,7 @@ module.exports = {
   extends: [
     'molindo/typescript',
     'molindo/react',
+    'molindo/tailwind',
     'plugin:@next/next/recommended'
   ]
 };